Add files via upload
The implementation of the Tsukamoto fuzzy inference system based on a website for employee candidate selection includes a variety of features designed to streamline the selection process. This system provides functionalities for inputting parameters, defining rules, establishing sets for each parameter, and managing job applicant periods. It also supports input from both applicants and employees.

Users can view graphical results for each job applicant, which facilitates a clear understanding of the evaluation outcomes. The defuzzification process produces a score ranging from 0 to 1, which is then sorted to identify the highest-ranking candidates. This sorting determines the acceptance or rejection status of each applicant for every recruitment period.

Additionally, the system includes features for posting announcements on the front-end of the application, a password change feature for login via PHPMailer, and various other functionalities to enhance user experience and system efficiency.